+++ cargo-embed/src/rttui/app.rs
@@ -4,8 +4,8 @@ use crossterm::{
     execute,
     terminal::{disable_raw_mode, enable_raw_mode, EnterAlternateScreen, LeaveAlternateScreen},
 };
+use probe_rs::rtt::RttChannel;
 use probe_rs::Core;
-use probe_rs_rtt::RttChannel;
 use std::{fmt::write, path::PathBuf, sync::mpsc::RecvTimeoutError};
 use std::{
     io::{Read, Seek, Write},
@@ -19,6 +19,8 @@ use tui::{
     widgets::{Block, Borders, List, ListItem, Paragraph, Tabs},
     Terminal,
 };
+
+use crate::DefmtInformation;
 
 use super::{
     channel::{ChannelState, DataFormat},
@@ -49,7 +51,7 @@ fn pull_channel<C: RttChannel>(channels: &mut Vec<C>, n: usize) -> Option<C> {
 
 impl App {
     pub fn new(
-        mut rtt: probe_rs_rtt::Rtt,
+        mut rtt: probe_rs::rtt::Rtt,
         config: &crate::config::Config,
         logname: String,
     ) -> Result<Self> {
@@ -154,10 +156,7 @@ impl App {
         None
     }
 
-    pub fn render(
-        &mut self,
-        defmt_state: &Option<(defmt_decoder::Table, Option<defmt_decoder::Locations>)>,
-    ) {
+    pub fn render(&mut self, defmt_state: Option<&DefmtInformation>) {
         let input = self.current_tab().input().to_owned();
         let has_down_channel = self.current_tab().has_down_channel();
         let scroll_offset = self.current_tab().scroll_offset();
@@ -289,16 +288,16 @@ impl App {
                                 ));
                             }
                             DataFormat::Defmt => {
-                                let (table, locs) = defmt_state.as_ref().expect(
+                                let defmt_state = defmt_state.as_ref().expect(
                                 "Running rtt in defmt mode but table or locations could not be loaded.",
                             );
-                                let mut stream_decoder = table.new_stream_decoder();
+                                let mut stream_decoder = defmt_state.table.new_stream_decoder();
                                 stream_decoder.received(&data);
                                 while let Ok(frame) = stream_decoder.decode()
                                 {
                                     // NOTE(`[]` indexing) all indices in `table` have already been
                                     // verified to exist in the `locs` map.
-                                    let loc = locs.as_ref().map(|locs| &locs[&frame.index()]);
+                                    let loc = defmt_state.location_information.as_ref().map(|locs| &locs[&frame.index()]);
 
                                     messages_wrapped.push(format!("{}", frame.display(false)));
                                     if let Some(loc) = loc {
@@ -479,11 +478,15 @@ impl App {
     /// Polls the RTT target for new data on all channels.
     ///
     /// # Errors
-    /// If getting the current time or formatting a timestamp fails,
+    /// If formatting a timestamp fails,
     /// this function will abort and return a [`time::Error`].
-    pub fn poll_rtt(&mut self, core: &mut Core) -> Result<(), time::Error> {
+    pub fn poll_rtt(
+        &mut self,
+        core: &mut Core,
+        offset: time::UtcOffset,
+    ) -> Result<(), time::Error> {
         for channel in self.tabs.iter_mut() {
-            channel.poll_rtt(core)?;
+            channel.poll_rtt(core, offset)?;
         }
 
         Ok(())