+++ probe-rs/src/config/target.rs
@@ -12,7 +12,7 @@ use crate::architecture::arm::sequences::{
     stm32h7::Stm32h7,
     ArmDebugSequence,
 };
-use crate::architecture::riscv::sequences::esp32c3::ESP32C3;
+use crate::architecture::riscv::sequences::{esp32c3::ESP32C3, esp32c6::ESP32C6};
 use crate::architecture::riscv::sequences::{DefaultRiscvSequence, RiscvDebugSequence};
 use crate::flashing::FlashLoader;
 use std::sync::Arc;
@@ -105,8 +105,11 @@ impl Target {
             tracing::warn!("Using custom sequence for LPC55S16/LPC55S69");
             debug_sequence = DebugSequence::Arm(LPC55S69::create());
         } else if chip.name.starts_with("esp32c3") {
-            tracing::warn!("Using custom sequence for ESP32c3");
+            tracing::warn!("Using custom sequence for ESP32C3");
             debug_sequence = DebugSequence::Riscv(ESP32C3::create());
+        } else if chip.name.starts_with("esp32c6") {
+            tracing::warn!("Using custom sequence for ESP32C6");
+            debug_sequence = DebugSequence::Riscv(ESP32C6::create());
         } else if chip.name.starts_with("nRF5340") {
             tracing::warn!("Using custom sequence for nRF5340");
             debug_sequence = DebugSequence::Arm(Nrf5340::create());