+++ probe-rs/src/flashing/progress.rs
@@ -1,5 +1,5 @@
 use super::FlashLayout;
-use std::time::Duration;
+use std::{sync::Arc, time::Duration};
 
 /// A structure to manage the flashing procedure progress reporting.
 ///
@@ -14,15 +14,16 @@ use std::time::Duration;
 /// // Print events
 /// let progress = FlashProgress::new(|event| println!("Event: {:#?}", event));
 /// ```
+#[derive(Clone)]
 pub struct FlashProgress {
-    handler: Box<dyn Fn(ProgressEvent)>,
+    handler: Arc<dyn Fn(ProgressEvent)>,
 }
 
 impl FlashProgress {
     /// Create a new `FlashProgress` structure with a given `handler` to be called on events.
     pub fn new(handler: impl Fn(ProgressEvent) + 'static) -> Self {
         Self {
-            handler: Box::new(handler),
+            handler: Arc::new(handler),
         }
     }
 
@@ -95,6 +96,11 @@ impl FlashProgress {
     pub(super) fn finished_filling(&self) {
         self.emit(ProgressEvent::FinishedFilling);
     }
+
+    #[cfg(feature = "rtt")]
+    pub(super) fn message(&self, message: String) {
+        self.emit(ProgressEvent::DiagnosticMessage { message });
+    }
 }
 
 /// Possible events during the flashing process.
@@ -164,4 +170,9 @@ pub enum ProgressEvent {
     FailedProgramming,
     /// Programming of the flash has finished successfully.
     FinishedProgramming,
+    /// a message was received from the algo.
+    DiagnosticMessage {
+        /// The message that was emitted.
+        message: String,
+    },
 }