+++ cargo-flash/CHANGELOG.md
@@ -5,23 +5,51 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Log messages are printed to stderr, not stdout.
+
+### Fixed
+
+- If the output was redirect into a file, some log messages were not printed.
+
+## [0.16.0]
+
+Released 2023-01-29
+
+cargo-flash binary is unchanged, version number is increased to keep in sync with other
+probe-rs packages.
+
+## [0.15.0]
+
+Released 2023-01-28
+### Changed
+
+- Updated probe-rs to 0.15.0.
+
 ## [0.14.2]
 
 Released 2023-01-18
+
+### Changed
+
+- Updated probe-rs to 0.14.2.
 
 ## [0.14.1]
 
 Released 2023-01-14
 
+### Changed
+
+- Updated probe-rs to 0.14.1.
+
 ## [0.14.0]
 
 Released 2023-01-13
 
-### Added
-
 ### Changed
 
-### Fixed
+- Updated probe-rs to 0.14.0.
 
 ## [0.13.0]
 
@@ -177,10 +205,12 @@ Improved flashing for `cargo-flash` considering speed and useability.
 
 - Introduce cargo-flash which can automatically build & flash the target elf file.
 
-[unreleased]: https://github.com/probe-rs/cargo-flash/compare/v0.14.2...master
-[v0.14.2]: https://github.com/probe-rs/cargo-flash/compare/v0.14.1...v0.14.2
-[v0.14.1]: https://github.com/probe-rs/cargo-flash/compare/v0.14.0...v0.14.1
-[v0.14.0]: https://github.com/probe-rs/cargo-flash/compare/v0.13.0...v0.14.0
+[unreleased]: https://github.com/probe-rs/cargo-flash/compare/v0.16.0...master
+[0.16.0]: https://github.com/probe-rs/cargo-flash/compare/v0.15.0...v0.16.0
+[0.15.0]: https://github.com/probe-rs/cargo-flash/compare/v0.14.2...v0.15.0
+[0.14.2]: https://github.com/probe-rs/cargo-flash/compare/v0.14.1...v0.14.2
+[0.14.1]: https://github.com/probe-rs/cargo-flash/compare/v0.14.0...v0.14.1
+[0.14.0]: https://github.com/probe-rs/cargo-flash/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/probe-rs/cargo-flash/releases/tag/v0.12.1..v0.13.0
 [0.12.1]: https://github.com/probe-rs/cargo-flash/releases/tag/v0.12.0..v0.12.1
 [0.12.0]: https://github.com/probe-rs/cargo-flash/releases/tag/v0.11.0..v0.12.0